Add bootcount sysfs to PCF8563 RTC driver
Driver now supports a file in the sysfs filesystem.

pcf8563_bootcount: USed to read and clear the bootcount register.

Writing to this will clear the Hour Alarm register (0x0A)

Minute Alarm register (0x09) is contains 0xBC signature to state
that the values stored in the alarm registers are invalid and used
for bootloader bootcount instead.