Update bootcount sysfs in PCF8563 RTC driver
Previously with the original U-Boot implementation of the i2c/RTC
bootcounter; only the signature (0xBC) was validated to query and
obtain the bootcount stored in the next register.  This caused
problems when the bootcount register was modified unknownly which
therefore reported an incorrect bootcount.

A CRC7 checksum is therefore performed upon the signature and
bootcounter to ensure the data is valid.

bootcount returns -1 if signature and/or CRC are invalid.

Writing to this will clear the count register (0x0B).