Reset RTC time to 01/01/1970 00:00:00 if PCF8563 year>=2038
A workaround solution to handle values stored in the PCF8563 RTC
which has its years' register to be greater than 0x38 (bcd) and
its century bit to be 0 (20xx) i.e. year >= 2038.

Due to the 32bit kernel limitation of storing the max possible
epoch time (2^32) we simply just reset the time stored in the RTC
to 01/01/1970 00:00:00 if this condition is met.

If in 2038 and we are still using a 32bit kernel with this exact
RTC driver; then we would need to revert this change.