Fix to RTC PCF8563 reset time if year >= 2038
Previously checked if internal year count was greater than 2038
and the century polarity was 20xx when obtaining the initial date
time from the RTC registers, the date time would be reset to epoch
time 0 if this was true.

Due to the unknown states of the registers simply checking whether
the year was >= 138 (due to all year dates <70 being treated as
year+=100), this ensures that the time reset fix is more reliable.